apps/bonde/requirements.txt
djangorestframework
django-appconf
django-filter
markdown
requests